move get_updated_options
