+++ packages/devextreme/js/__internal/core/r1/dom_component_wrapper.tsx
@@ -5,8 +5,8 @@ import {
   hasTemplate, InfernoComponent, InfernoEffect, renderTemplate,
 } from '@devextreme/runtime/inferno';
 import type { ComponentClass } from '@js/core/dom_component';
-import { getUpdatedOptions } from '@js/renovation/ui/common/utils/get_updated_options';
 import type { DisposeEffectReturn } from '@js/renovation/utils/effect_return';
+import { getUpdatedOptions } from '@ts/core/r1/utils/get_updated_options';
 import type { RefObject } from 'inferno';
 import { createRef } from 'inferno';
