+++ packages/devextreme/js/__internal/core/r1/utils/get_updated_options.test.ts
@@ -1,4 +1,4 @@
-import { getUpdatedOptions } from '../get_updated_options';
+import { getUpdatedOptions } from './get_updated_options';
 
 describe('get_updated_options', () => {
   class DummyDataSource { dummy = true; }
